requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
cryptography==45.0.3
distro==1.9.0
Django==5.2.1
django-allauth==65.9.0
django-browser-reload==1.18.0
django-extensions==4.1
django-tailwind==4.0.1
google-auth==2.40.2
google-genai==1.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
openai==1.82.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
